Change About menu windows from Alert to Stage
JavaFX Alerts behave strange on some Linux systems and are not automatically resized to allow for the entire alert's contentText to be displayed. This makes the text unreadable on desktop environments that don't allow resizing of modal dialogs. Instead, placing this info in a new Stage object will ensure that the user can resize this "pop-up" if it is not sized correctly.

Signed-off-by: dustinkredmond <dustinkredmond@gmail.com>